Backport: Add GTK_NO_TAB_SCROLL to turn off tab scrolling
MATE's Pluma text editor document tabs were scrolling twice. Fixes #47.

Strangely, not all apps that reimplemented this feature are affected. It
might only be occuring on certain types of Notebook controls. This
change allows users to turn it off if need be, such as prepending the
variable in the application launcher, like so:

GTK_NO_TAB_SCROLL=1 pluma %U