Fix: Some non-Wacom graphics/drawing tablet styli were being recognised as GDK_SOURCE_TOUCHSCREEN input sources.
Some digitizers (like the Huion H1060P) get enumerated by the 'wacom'
driver in X as being of type 'stylus', and the word 'stylus' is appended
to their names (as seen by xinput).

GDK decides input source types by looking at device name strings when other
techniques have failed to definitively provide an answer.

The only words GDK looks for to detect a stylus input device are 'wacom'
and 'pen'.

The name of the Huion H1060P's stylus (for examplee) is
'HID 256c:006d stylus' according to XInput, which matches neither the
words  'wacom' nor 'pen'.

This patch simply adds a match for the word 'stylus' to fix this and
correctly recognise styli as being of type GDK_SOURCE_PEN.