Attempt packaging for Ubuntu 24.04
Some changes are required in order to build on Launchpad for the
upcoming Ubuntu 24.04 LTS.

Not sure why, but even though it's based on the original gtk3 packaging,
we can't actually reproduce the build on Launchpad because it's stuck on
a "dependency wait" for dpkg-dev (>= 1.22.5). Also, at time of commit,
3.24.41 is being built even though 3.24.40 is currently in the
(pre-beta) archive.

It fails to build:
https://launchpad.net/~lah7/+archive/ubuntu/gtk3-classic/+build/27957926