Ubuntu packaging script: Change changelog message
Ubuntu packages (at this time) typically get built once and rarely
rebuilt again if there are new patches in master. This is due to
different GTK versions and the complexity of the packaging.

Add more emphasis on the commit hash to be clear that we take an
existing GTK 3 Ubuntu package (upstream) and add our patches on top.