Add script to refresh patches
Useful in any scenario where fuzz patches are not accepted.

At the moment, there's no issues with fuzzy patches.

I like that some of the patches haven't been touched in years, it helps
tell the story of how they came to be. Some even reveal the GTK source
folder they were originally patched from.

Running it would cause unnecessary clutter in the Git history, and be a
game of "cat-and-mouse" keeping it up-to-date, so it's here just in case
it becomes a necessity.