Fix missing dialog icons in GTK 3.24.48
Upstream caused a patch conflict:
https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7587

The changes in #114 (9129a635dfbf) to fix the patch were not 100%
complete. The icon variable was set outside [update_accessible_name] of
the function [setup_type] that actually sets this icon. Even though
repeating the case select is a bit like code duplication, it works and
fixes the regression.