PKGBUILD: Use wildcard to source all patch files
To ease maintenance and reduce friction for contributors
adding new patches who may not be using Arch Linux.

This does not impact the integrity checks. This 'master'
branch must continue to always have up-to-date checksums.

I did consider to 'SKIP' checksums, but that'll be a bad idea.
We need integrity checks for remote sources, and we can't
'SKIP' just the patch files. Plus, release assets are built
from this PKGBUILD.

This does NOT affect the AUR version of the package.