CI: Check PKGBUILD sha256sums in a separate job
To reduce friction when there are contributions from users who may not
be using Arch Linux.

Always build the package regardless (for pull requests) to speed up
testing by skipping an integrity check.

I'd love to be able to automatically commit the changes and restart the
workflow, but this can't be done in someone's fork for security reasons
using GitHub Actions alone.